refactor: Address PR #30 review feedback
- Move AccountBuffer and spawn_buffer_task to account_buffer.rs
- Extract shared forward_message() to deduplicate forwarding logic
  between buffer task and PluginInner::send_message
- Use if-let-else instead of match in upsert for readability
- Default account_buffer to Some (enabled by default)
- Remove is_startup guard (startup accounts use snapshot channel)
